requirements.txt
packaging==24.0
pillow==10.3.0
pip==24.0
pytesseract==0.3.10
packaging==24.0
streamlit==1.22.0
opencv-python==4.5.5.64
keras-ocr==0.8.8
easyocr==1.5.0
tensorflow==2.14.1
transformers==4.18.0
torch==2.0.1
